requirements.txt
mouse
pyautogui
openai
asyncio
pytest-playwright